pyproject.toml
[tool.poetry]
name = "no-more-forwards"
version = "0.1.0"
description = ""
authors = ["vile <111662603+vile@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
